+++ CHANGES.md
@@ -14,8 +14,8 @@
 ## 0.7.1 / 2019-January-??
 
 * Routes
-  - [Todo] Add utility fn `calfpath.route/prewalk-routes`
-  - [Todo] Fix reporting "URI templates" in routes fallback-400 handler
+  - Add utility fn `calfpath.route/prewalk-routes`
+  - Fix reporting "URI templates" in routes fallback-400 handler
     - Introduce `:full-uri` kwarg in `calfpath.route/compile-routes` as reference key